added embedding functions and dataset

+++ embeddings.py
@@ -4,9 +4,5 @@ import torch
 def gene_embedding(id):
     return torch.ones(20)
 
-
 def disease_embedding(id):
-    return torch.ones(20)
-
-
-
+    return torch.ones(20, dtype=torch.float32)
